# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Redaction review API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  pending: 'pending',
  in_review: 'in_review',
  completed: 'completed',
} as const;

export interface Document {
  id: string;
  title: string;
  status: DocumentStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  totalRedactions: number;
  pendingCount: number;
  confirmedCount: number;
  rejectedCount: number;
  userAddedCount: number;
}

export type DocumentDetailStatus = typeof DocumentDetailStatus[keyof typeof DocumentDetailStatus];


export const DocumentDetailStatus = {
  pending: 'pending',
  in_review: 'in_review',
  completed: 'completed',
} as const;

export type RedactionCategory = typeof RedactionCategory[keyof typeof RedactionCategory];


export const RedactionCategory = {
  name: 'name',
  phone: 'phone',
  email: 'email',
  address: 'address',
  ssn: 'ssn',
  dob: 'dob',
  financial: 'financial',
  medical: 'medical',
  organization: 'organization',
  other: 'other',
} as const;

export type RedactionStatus = typeof RedactionStatus[keyof typeof RedactionStatus];


export const RedactionStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rejected: 'rejected',
  user_added: 'user_added',
} as const;

export type RedactionSource = typeof RedactionSource[keyof typeof RedactionSource];


export const RedactionSource = {
  ai: 'ai',
  user: 'user',
} as const;

export interface Redaction {
  id: string;
  documentId: string;
  startOffset: number;
  endOffset: number;
  text: string;
  category: RedactionCategory;
  confidence: number;
  status: RedactionStatus;
  source: RedactionSource;
  /** @nullable */
  note?: string | null;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface DocumentDetail {
  id: string;
  title: string;
  status: DocumentDetailStatus;
  content: string;
  redactions: Redaction[];
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface DocumentInput {
  title: string;
  content: string;
}

export type RedactionInputCategory = typeof RedactionInputCategory[keyof typeof RedactionInputCategory];


export const RedactionInputCategory = {
  name: 'name',
  phone: 'phone',
  email: 'email',
  address: 'address',
  ssn: 'ssn',
  dob: 'dob',
  financial: 'financial',
  medical: 'medical',
  organization: 'organization',
  other: 'other',
} as const;

export interface RedactionInput {
  startOffset: number;
  endOffset: number;
  text: string;
  category: RedactionInputCategory;
  note?: string;
}

export type RedactionUpdateStatus = typeof RedactionUpdateStatus[keyof typeof RedactionUpdateStatus];


export const RedactionUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rejected: 'rejected',
  user_added: 'user_added',
} as const;

export type RedactionUpdateCategory = typeof RedactionUpdateCategory[keyof typeof RedactionUpdateCategory];


export const RedactionUpdateCategory = {
  name: 'name',
  phone: 'phone',
  email: 'email',
  address: 'address',
  ssn: 'ssn',
  dob: 'dob',
  financial: 'financial',
  medical: 'medical',
  organization: 'organization',
  other: 'other',
} as const;

export interface RedactionUpdate {
  status?: RedactionUpdateStatus;
  note?: string;
  category?: RedactionUpdateCategory;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ReviewSummary {
  documentId: string;
  totalRedactions: number;
  confirmedCount: number;
  rejectedCount: number;
  userAddedCount: number;
  pendingCount: number;
  riskScore: number;
  completionPercent: number;
  categoryBreakdown: CategoryCount[];
}

export type SuspiciousSpanRiskLevel = typeof SuspiciousSpanRiskLevel[keyof typeof SuspiciousSpanRiskLevel];


export const SuspiciousSpanRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface SuspiciousSpan {
  startOffset: number;
  endOffset: number;
  text: string;
  reason: string;
  riskLevel: SuspiciousSpanRiskLevel;
}