# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Redaction review API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Document operations
  - name: redactions
    description: Redaction operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document with all redactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document with redactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /documents/{id}/summary:
    get:
      operationId: getDocumentSummary
      tags: [documents]
      summary: Get privacy review summary for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Privacy summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewSummary"

  /documents/{id}/complete:
    post:
      operationId: completeReview
      tags: [documents]
      summary: Mark review as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Review completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}/redactions:
    get:
      operationId: listRedactions
      tags: [redactions]
      summary: Get all redactions for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of redactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Redaction"
    post:
      operationId: createRedaction
      tags: [redactions]
      summary: Add a new redaction (user-marked missed PII)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redaction"

  /documents/{id}/redactions/{redactionId}:
    patch:
      operationId: updateRedaction
      tags: [redactions]
      summary: Update a redaction status (confirm, reject, or add note)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: redactionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedactionUpdate"
      responses:
        "200":
          description: Updated redaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redaction"
    delete:
      operationId: deleteRedaction
      tags: [redactions]
      summary: Delete a user-added redaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: redactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /documents/{id}/suspicious:
    get:
      operationId: getSuspiciousText
      tags: [redactions]
      summary: Get unredacted text spans that look like PII (missed detections)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of suspicious unredacted spans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuspiciousSpan"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Document:
      type: object
      required: [id, title, status, createdAt, totalRedactions, pendingCount, confirmedCount, rejectedCount, userAddedCount]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [pending, in_review, completed]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
        totalRedactions:
          type: integer
        pendingCount:
          type: integer
        confirmedCount:
          type: integer
        rejectedCount:
          type: integer
        userAddedCount:
          type: integer

    DocumentDetail:
      type: object
      required: [id, title, status, content, redactions, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [pending, in_review, completed]
        content:
          type: string
        redactions:
          type: array
          items:
            $ref: "#/components/schemas/Redaction"
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]

    DocumentInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string

    Redaction:
      type: object
      required: [id, documentId, startOffset, endOffset, text, category, confidence, status, source]
      properties:
        id:
          type: string
        documentId:
          type: string
        startOffset:
          type: integer
        endOffset:
          type: integer
        text:
          type: string
        category:
          type: string
          enum: [name, phone, email, address, ssn, dob, financial, medical, organization, other]
        confidence:
          type: number
        status:
          type: string
          enum: [pending, confirmed, rejected, user_added]
        source:
          type: string
          enum: [ai, user]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]

    RedactionInput:
      type: object
      required: [startOffset, endOffset, text, category]
      properties:
        startOffset:
          type: integer
        endOffset:
          type: integer
        text:
          type: string
        category:
          type: string
          enum: [name, phone, email, address, ssn, dob, financial, medical, organization, other]
        note:
          type: string

    RedactionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, rejected, user_added]
        note:
          type: string
        category:
          type: string
          enum: [name, phone, email, address, ssn, dob, financial, medical, organization, other]

    ReviewSummary:
      type: object
      required: [documentId, totalRedactions, confirmedCount, rejectedCount, userAddedCount, pendingCount, riskScore, categoryBreakdown, completionPercent]
      properties:
        documentId:
          type: string
        totalRedactions:
          type: integer
        confirmedCount:
          type: integer
        rejectedCount:
          type: integer
        userAddedCount:
          type: integer
        pendingCount:
          type: integer
        riskScore:
          type: integer
        completionPercent:
          type: number
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"

    CategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer

    SuspiciousSpan:
      type: object
      required: [startOffset, endOffset, text, reason, riskLevel]
      properties:
        startOffset:
          type: integer
        endOffset:
          type: integer
        text:
          type: string
        reason:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]