# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Redaction review API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from './documentStatus';

export interface Document {
  id: string;
  title: string;
  status: DocumentStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  totalRedactions: number;
  pendingCount: number;
  confirmedCount: number;
  rejectedCount: number;
  userAddedCount: number;
}