# lib/api-zod/src/generated/types/documentDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Redaction review API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentDetailStatus } from './documentDetailStatus';
import type { Redaction } from './redaction';

export interface DocumentDetail {
  id: string;
  title: string;
  status: DocumentDetailStatus;
  content: string;
  redactions: Redaction[];
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}