# lib/api-zod/src/generated/types/redaction.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Redaction review API
 * OpenAPI spec version: 0.1.0
 */
import type { RedactionCategory } from './redactionCategory';
import type { RedactionSource } from './redactionSource';
import type { RedactionStatus } from './redactionStatus';

export interface Redaction {
  id: string;
  documentId: string;
  startOffset: number;
  endOffset: number;
  text: string;
  category: RedactionCategory;
  confidence: number;
  status: RedactionStatus;
  source: RedactionSource;
  /** @nullable */
  note?: string | null;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
}