# lib/api-zod/src/generated/types/redactionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Redaction review API
 * OpenAPI spec version: 0.1.0
 */
import type { RedactionInputCategory } from './redactionInputCategory';

export interface RedactionInput {
  startOffset: number;
  endOffset: number;
  text: string;
  category: RedactionInputCategory;
  note?: string;
}