# lib/api-zod/src/generated/types/redactionUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Redaction review API
 * OpenAPI spec version: 0.1.0
 */
import type { RedactionUpdateCategory } from './redactionUpdateCategory';
import type { RedactionUpdateStatus } from './redactionUpdateStatus';

export interface RedactionUpdate {
  status?: RedactionUpdateStatus;
  note?: string;
  category?: RedactionUpdateCategory;
}