# lib/api-zod/src/generated/types/reviewSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Redaction review API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';

export interface ReviewSummary {
  documentId: string;
  totalRedactions: number;
  confirmedCount: number;
  rejectedCount: number;
  userAddedCount: number;
  pendingCount: number;
  riskScore: number;
  completionPercent: number;
  categoryBreakdown: CategoryCount[];
}