# lib/api-zod/src/generated/types/suspiciousSpan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Redaction review API
 * OpenAPI spec version: 0.1.0
 */
import type { SuspiciousSpanRiskLevel } from './suspiciousSpanRiskLevel';

export interface SuspiciousSpan {
  startOffset: number;
  endOffset: number;
  text: string;
  reason: string;
  riskLevel: SuspiciousSpanRiskLevel;
}